add owner permissions to permission output
The owner was previously not returned, because they do not have a dedicated "permission" ID / entry. Now we append a manual record with the owner's GUID and an NA_character_ ID, so that it is easy to determine who the owner is without changing the API signature. We also add a parameter to opt out of this behavior, since it is inconsistent with the API.

NOTE that we leave the default to add_owner = FALSE for the lower-level / internal functions so that it maps to the API more directly.