allow opting out of the empty object expectation
"Most" Connect endpoints expect an empty map. R does not provide such a mechanism AFAICT and presumes that an empty list is a list (unless perhaps you shove names along). However, some (i.e. environments) expect an array, so we have to provide a mechanism to avoid the new default.