add tests to cover basic schedule behavior
Schedules are particularly tricky to test because time zones are wacky, formats
vary, and everything is relative (i.e. time elapses).  We also need more
coverage for the different schedule types.

Also switch the debug display to be a bit more helpful (showing the actual URL
requested, etc.). It is unclear if this will cause problems for certain error
states (using the `res` variable instead of `req`).